tech!: Deprecate pipenv in favour of uv
pipenv is no longer a recommended python package and manager
version. uv does more, takes less time, and is easier to live with.

This commit removes pipenv in favour of uv, including updating the
runfile.

Migration: `./run sync` installs packages, and `./run dev` runs a
development environment as before.

BREAKING CHANGE: Requires uv to be present on the machine or in the
container, and any additional packages in your Pipenv environment
will need to be reinstalled via `uv add`.